zmarkdown: rename modules to 'modules'
